version 0.95

+++ src/random_wrap.c
@@ -6,6 +6,6 @@
  */
 void F77_SUB(rndstart)(void) { GetRNGstate(); }
 void F77_SUB(rndend)(void) { PutRNGstate(); }
-double F77_SUB(random)() { return unif_rand(); }
+float F77_SUB(random)() { return unif_rand(); }
